RVM-1062 : Convert int constants into offsets to fix 64-bit issues (original changes by Ian Rogers for MRP).
See https://github.com/codehaus/mrp/commit/d19a7989e5fb45e8a4542c6d97f880329a268767 for the original MRP commit.

This commit also adds a few tests for Offset.

+++ rvm/src/org/jikesrvm/ia32/OutOfLineMachineCode.java
@@ -262,9 +262,9 @@ public abstract class OutOfLineMachineCode implements BaselineConstants {
     ThreadLocalState.emitMoveRegToField(asm, fpOffset, SP); // establish base of new frame
     asm.emitPUSH_Imm(INVISIBLE_METHOD_ID);
     if (VM.BuildFor32Addr) {
-      asm.emitADD_Reg_Imm(SP, STACKFRAME_BODY_OFFSET);
+      asm.emitADD_Reg_Imm(SP, STACKFRAME_BODY_OFFSET.toInt());
     } else {
-      asm.emitADD_Reg_Imm_Quad(SP, STACKFRAME_BODY_OFFSET);
+      asm.emitADD_Reg_Imm_Quad(SP, STACKFRAME_BODY_OFFSET.toInt());
     }
 
     /* write parameters on stack
@@ -441,9 +441,9 @@ public abstract class OutOfLineMachineCode implements BaselineConstants {
     // NOTE: RVM callee has popped the params, so we can simply
     //       add back in the initial SP to FP delta to get SP to be a framepointer again!
     if (VM.BuildFor32Addr) {
-      asm.emitADD_Reg_Imm(SP, -STACKFRAME_BODY_OFFSET + WORDSIZE);
+      asm.emitADD_Reg_Imm(SP, -STACKFRAME_BODY_OFFSET.toInt() + WORDSIZE);
     } else {
-      asm.emitADD_Reg_Imm_Quad(SP, -STACKFRAME_BODY_OFFSET + WORDSIZE);
+      asm.emitADD_Reg_Imm_Quad(SP, -STACKFRAME_BODY_OFFSET.toInt() + WORDSIZE);
     }
     asm.emitPOP_RegDisp(TR, fpOffset);
